Flaskr is a blog application created with Flask
Users are able to register, log in, create posts, and edit or delete
their own posts. At this stage, the application is an exact copy of the
tutorial on Flask's documentation website. In the future, I will:
- Make the project installable
- Test coverage
- Deploy to production
- Keep developing the application.

To run this locally, use Python version 3.6 and newer, install venv,
create and activate a virtual environment, and within the activated
environment, install Flask.

This is the list of installed packages:

Package      Version
------------ -------
click        8.0.1
colorama     0.4.4
Flask        2.0.1
itsdangerous 2.0.1
Jinja2       3.0.1
MarkupSafe   2.0.1
pip          21.1.3
setuptools   56.0.0
Werkzeug     2.0.1